# api-laravel-main/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IFRI-UAC notation, API Documentation",
        "description": "IFRI-UAC notation, API Documentation",
        "contact": {
            "email": "contact@equipe-api.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.43.5:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Login",
                "description": "Returns user with access token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/login-anonymous": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Anonymous",
                "description": "Returns user with access token",
                "operationId": "loginAnonymous",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/users/generate-credentials": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Generate credentials",
                "description": "Returns users list",
                "operationId": "generateCredentials",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CredentialRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CredentialModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/edutiants/{classe_id}": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Get credentials by classe",
                "description": "Returns users list",
                "operationId": "getCredentials",
                "parameters": [
                    {
                        "name": "classe_id",
                        "in": "path",
                        "description": "Classe id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CredentialModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/users/role/{name}": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Get users by name",
                "description": "Returns users list",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserModel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification - Logout",
                "description": "Returns message",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AuthResponse": {
                "title": "Auth Response",
                "properties": {
                    "user": {
                        "title": "authentification"
                    },
                    "access_token": {
                        "title": "access_token",
                        "type": "string"
                    },
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthResource"
                }
            },
            "LoginRequest": {
                "title": "Login Request",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "admin1998@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "1993_188jjdjdjz"
                    }
                },
                "type": "object"
            },
            "LogoutResponse": {
                "title": "Logout Response",
                "properties": {
                    "message": {
                        "title": "message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LogoutResource"
                }
            },
...
            "UserModel": {
                "title": "User Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Identifier",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of user",
                        "type": "string",
                        "example": "Robert"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "admin1998@gmail.com"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of user",
                        "type": "string",
                        "example": "ETUDIANT"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "CredentialModel": {
                "title": "Credential Model",
                "properties": {
                    "qrcode": {
                        "title": "qrcode",
                        "description": "CodeQr value",
                        "type": "string",
                        "example": "Koakdkvardnh19373JDJSJK72736YYSHJHJHDUY3676UJHD267ESJS"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Credential"
                }
            },
            "CredentialRequest": {
                "title": "Generate credentials Request",
                "required": [
                    "classe_id",
                    "effectif"
                ],
                "properties": {
                    "classe_id": {
                        "title": "classe_id",
                        "description": "Classe id",
                        "type": "integer",
                        "example": "2"
                    },
                    "effectif": {
                        "title": "effectif",
                        "description": "Effectif of the classe",
                        "type": "integer",
                        "example": "57"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "The value of access_token received from login: Ex: Bearer access_token_value",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}